remove obsolete utility
