+++ tests/org.eclipse.n4js.hlc.tests/src/org/eclipse/n4js/hlc/tests/TargetPlatformConfigurationsTest.java
@@ -10,12 +10,16 @@
  */
 package org.eclipse.n4js.hlc.tests;
 
+import static com.google.common.base.Preconditions.checkState;
+import static org.eclipse.n4js.utils.io.FileUtils.createDirectory;
+import static org.eclipse.n4js.utils.io.FileUtils.createTempDirectory;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
 
 import java.io.File;
 import java.io.IOException;
+import java.nio.file.Path;
 
 import org.eclipse.n4js.hlc.base.BuildType;
 import org.eclipse.n4js.hlc.base.ErrorExitCode;
@@ -23,6 +27,7 @@ import org.eclipse.n4js.hlc.base.ExitCodeException;
 import org.eclipse.n4js.hlc.base.N4jscBase;
 import org.eclipse.n4js.hlc.base.SuccessExitStatus;
 import org.eclipse.n4js.utils.io.FileDeleter;
+import org.eclipse.n4js.utils.io.FileUtils;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
@@ -33,16 +38,37 @@ import org.junit.Test;
 public class TargetPlatformConfigurationsTest extends AbstractN4jscTest {
 	File workspace;
 
+	private final TargetPlatformFiles platformFiles = new TargetPlatformFiles();
+
 	/** Prepare workspace. */
 	@Before
 	public void setupWorkspace() throws IOException {
+		setupExternals(platformFiles, description.getMethodName());
 		workspace = setupWorkspace("external_with_n4jsd_tpt");
 	}
 
 	/** Delete workspace. */
 	@After
 	public void deleteWorkspace() throws IOException {
+		cleanupExternals(platformFiles);
 		FileDeleter.delete(workspace.toPath(), true);
+	}
+
+	/** Initializes the target platform install location. */
+	public static void setupExternals(TargetPlatformFiles platformFiles, String tmpPrefix) {
+		checkState(null == platformFiles.targetPlatformInstallLocation);
+
+		final Path tempRoot = createTempDirectory("hlcTest-time-" + System.currentTimeMillis());
+		platformFiles.root = tempRoot.toFile();
+		platformFiles.targetPlatformInstallLocation = createDirectory(tempRoot, tmpPrefix).toFile();
+	}
+
+	/** Cleans up the target platform install location. */
+	public static void cleanupExternals(TargetPlatformFiles platformFiles) {
+		if (null != platformFiles.targetPlatformInstallLocation) {
+			FileUtils.deleteFileOrFolder(platformFiles.targetPlatformInstallLocation);
+			platformFiles.targetPlatformInstallLocation = null;
+		}
 	}
 
 	/**
@@ -56,18 +82,18 @@ public class TargetPlatformConfigurationsTest extends AbstractN4jscTest {
 		final String wsRoot = workspace.getAbsolutePath().toString();
 
 		// force creating install location
-		FileDeleter.delete(getTargetPlatformInstallLocation());
+		FileDeleter.delete(platformFiles.targetPlatformInstallLocation);
 
 		final String[] args = {
 				"--installMissingDependencies",
-				"--targetPlatformInstallLocation", getTargetPlatformInstallLocation().getAbsolutePath(),
+				"--targetPlatformInstallLocation", platformFiles.targetPlatformInstallLocation.getAbsolutePath(),
 				"--verbose",
 				"--projectlocations", wsRoot,
 				"--buildType", BuildType.allprojects.toString()
 		};
 		SuccessExitStatus status = new N4jscBase().doMain(args);
 		assertEquals("Should exit with success", SuccessExitStatus.INSTANCE.code, status.code);
-		assertTrue("install location was not created", getTargetPlatformInstallLocation().exists());
+		assertTrue("install location was not created", platformFiles.targetPlatformInstallLocation.exists());
 	}
 
 	/**
@@ -81,13 +107,13 @@ public class TargetPlatformConfigurationsTest extends AbstractN4jscTest {
 		final String wsRoot = workspace.getAbsolutePath().toString();
 
 		// force creating install location
-		File testFile = new File(getTargetPlatformInstallLocation(), "tst.txt");
+		File testFile = new File(platformFiles.targetPlatformInstallLocation, "tst.txt");
 		testFile.createNewFile();
 		assertTrue("setup error, test file should exist yet at " + testFile.getAbsolutePath(), testFile.exists());
 
 		final String[] args = {
 				"--clean",
-				"--targetPlatformInstallLocation", getTargetPlatformInstallLocation().getAbsolutePath(),
+				"--targetPlatformInstallLocation", platformFiles.targetPlatformInstallLocation.getAbsolutePath(),
 				"--verbose",
 				"--projectlocations", wsRoot,
 				"--buildType", BuildType.allprojects.toString()
@@ -107,7 +133,7 @@ public class TargetPlatformConfigurationsTest extends AbstractN4jscTest {
 
 		final String[] args = {
 				// "--targetPlatformFile", getTargetPlatformFile().getAbsolutePath(),
-				"--targetPlatformInstallLocation", getTargetPlatformInstallLocation().getAbsolutePath(),
+				"--targetPlatformInstallLocation", platformFiles.targetPlatformInstallLocation.getAbsolutePath(),
 				"--verbose",
 				"--projectlocations", wsRoot,
 				"--buildType", BuildType.allprojects.toString()
@@ -128,25 +154,25 @@ public class TargetPlatformConfigurationsTest extends AbstractN4jscTest {
 
 		final String[] argsInstall = {
 				"--installMissingDependencies",
-				"--targetPlatformInstallLocation", getTargetPlatformInstallLocation().getAbsolutePath(),
+				"--targetPlatformInstallLocation", platformFiles.targetPlatformInstallLocation.getAbsolutePath(),
 				"--verbose",
 				"--projectlocations", wsRoot,
 				"--buildType", BuildType.allprojects.toString()
 		};
 		SuccessExitStatus statusInstall = new N4jscBase().doMain(argsInstall);
 		assertEquals("Should exit with success", SuccessExitStatus.INSTANCE.code, statusInstall.code);
-		assertTrue("install location was not created", getTargetPlatformInstallLocation().exists());
+		assertTrue("install location was not created", platformFiles.targetPlatformInstallLocation.exists());
 
 		final String[] argsSkipInstall = {
 				// "--installMissingDependencies",
-				"--targetPlatformInstallLocation", getTargetPlatformInstallLocation().getAbsolutePath(),
+				"--targetPlatformInstallLocation", platformFiles.targetPlatformInstallLocation.getAbsolutePath(),
 				"--verbose",
 				"--projectlocations", wsRoot,
 				"--buildType", BuildType.allprojects.toString()
 		};
 		SuccessExitStatus statusSkipInstall = new N4jscBase().doMain(argsSkipInstall);
 		assertEquals("Should exit with success", SuccessExitStatus.INSTANCE.code, statusSkipInstall.code);
-		assertTrue("install location still exists", getTargetPlatformInstallLocation().exists());
+		assertTrue("install location still exists", platformFiles.targetPlatformInstallLocation.exists());
 
 	}
 }
